Make View.alpha float and optimize View colorTransform to use floats and include only multiplicative tinting

+++ korge-sandbox/src/commonMain/kotlin/samples/minesweeper/RandomLight.kt
@@ -34,7 +34,7 @@ class RandomLight(
 		incs = random[0.5, 2.0]
 		excx = random[0.7, 1.3]
 		excy = random[0.7, 1.3]
-		alpha = 0.1
+		alpha = 0.1f
 
 		while (true) {
 			rotation += inca.degrees
@@ -47,9 +47,9 @@ class RandomLight(
             // Check if a light sphere collided with another one
 			// The default collision system is inner circle
 			if (this.collision<RandomLight>() != null) {
-				alpha = (alpha + 0.01).coerceIn(0.1, 0.8)
+				alpha = (alpha + 0.01).coerceIn(0.1, 0.8).toFloat()
 			} else {
-				alpha = (alpha - 0.05).coerceIn(0.1, 0.8)
+				alpha = (alpha - 0.05).coerceIn(0.1, 0.8).toFloat()
 			}
 
 			frame()
